Introduce holdEmitter
This will let us add some delay between send retries, which is probably
a good idea to e.g. protect against InfluxDB failures. Currently, there
is no limit to the number of in-flight batches in the holdEmitter. In
practices, the rate is limited by the upstream emitter, so that's
probably fine.