Fix racy Batcher test
If the batcher starts its run() method before we Close() it, then it'll
try to emit a batch before closing, which will timeout if this channel
is unbuffered.

Broadly speaking, we should add test coverage that the Batcher properly
Emits() before closing, but that's not actually the case right now.